+++ CHANGES.md
@@ -3,13 +3,15 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 # 4.0.1
-- Added the ability to assert that signals did or did not fire.  By calling `watch_signals` Gut will monitor any signals emitted by the object.  You can then later make assertions about whether or not a signal was emitted or whether it emitted a desired number of times.  You can also verify that the signal was emitted with the right parameters.  The following methods were added for this feature, see the README for more information about them.
+- Added the ability to assert that signals did or did not fire.  By calling `watch_signals` Gut will monitor any signals emitted by the object.  You can then later make assertions about whether or not a signal was emitted or whether it emitted a desired number of times, and even if it was emitted and passed specific parameters.  You can also verify that the signal was emitted with the right parameters.  The following methods were added for this feature, see the README for more information about them.
 	- `watch_signals`
 	- `assert_signal_emitted`
 	- `assert_signal_not_emitted`
+	- `assert_signal_emitted_with_parameters`
 	- `assert_signal_emit_count`
 	- `assert_has_signal`
 	- `get_signal_emit_count`
+	- `get_signal_parameters`
 
 # 4.0.0
 ### Breaking Changes from 3.0.x and earlier